requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
uritemplate==4.1.1
